backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
Pillow

